refactor: add get-webpack-assets function and mock data
