+++ packages/scrollable-image/src/components/test-component.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import ScrollHorizontal from './scroll-horizontal'
 import styled from 'styled-components'
+import { mockImgSrc, mockImgSrcs } from '../../dev/mock-data'
 
 const FullPage1 = styled.div`
   position: relative;
@@ -16,28 +17,16 @@ const FullPage2 = styled.div`
   background: green;
 `
 
-const imgSrc = [
-  'https://static01.nyt.com/newsgraphics/2016/08/14/men-100-meters-bolt-horizontal/09c0dfe010da583c01f23709a11f6153e10cbb7b/bolt-100m-race-a3698x450.jpg',
-]
-
-const imgSrcs = Array.apply(null, Array(5)).map(() => {
-  const min = 50
-  const max = 500
-  const height = Math.floor(Math.random() * (max - min + 1)) + min
-  const width = Math.floor(Math.random() * (max - height + 1)) + height
-  return `https://picsum.photos/${width}/${height}`
-})
-
 export default class TestComponent extends React.PureComponent {
   render() {
     const EmbeddedInArticle = (
       <>
         <FullPage1 />
         <FullPage1 />
-        <ScrollHorizontal imgSrc={imgSrc} />
+        <ScrollHorizontal imgSrc={mockImgSrc} />
         <FullPage2 />
         <FullPage1 />
-        <ScrollHorizontal imgSrc={imgSrcs} lazyload />
+        <ScrollHorizontal imgSrc={mockImgSrcs} lazyload />
         <FullPage2 />
       </>
     )
